Collections/BaseDictionaries.cs: * Added an optimized implementation of HashSetDictionary.AddRange() + Added MultiValueDictionaryBase.AddRange(IEnumerable<TKey>, TValue) to add a value to several keys * Added some parameter validation
Collections/BaseLists.cs:
* Simplified CollectionBase<T>.AddRange()

Collections/Caching.cs:
* Fixed bugs in the DictionaryCache.GetAllItems() method
* Renamed DictionaryCache.GetAllItems() to .GetAllEntries()
* Made DictionaryCache.Contains() always check but never update the expiration time of the item
* Made DictionaryCache.ContainsKey() remove the entry if it's expired

IO/BinaryReaderWriter.cs:
* Minor simplification in BinaryReader.ReadEncodedInt32() extension

Transactions/documentation.xml:
Transactions/STM.cs:
Transactions/Tests/STM.cs:
+ Added STM.AddPostCommitAction and STMTransaction.AddPostCommitAction to allow you to register
  post-commit actions before the end of the transaction
* Fixed a bug in reading an uncommitted value when the current transaction appears to have been
  suppressed

Utilities/Linq.cs:
+ Added EnumerableExtensions.Coalesce() overrides for three or more sequences
+ Added EnumerableExtensions.MergeJoin() for merging sorted sequences
+ Added an IEnumerable<T>.WhereNotNull() extension for nullable value types
* Avoided duplicate lambda expressions (which would lead to duplicate generated functions)
* Made generator functions validate arguments eagerly

Utilities/TopologicalSort.cs:
* Replaced Converter<T,IEnumerable<T>> with Func<T,IEnumerable<T>> because it's not really
  converting the value
* Made generator functions validate arguments eagerly

Utilities/XmlUtility.cs:
* Added some parameter validation

+++ Collections/Tests/BloomFilter.cs
@@ -43,7 +43,7 @@ public class BloomFilterTest
             i++;
           } while(i < end);
         }, itemCount <= 100000 ? 1 : SystemInformation.GetAvailableCpuThreads());
-        float actualRate = (float)collisions / (itemCount*4);
+        double actualRate = (double)collisions / (itemCount*4);
 
         // now we'll apply a statistical significance test to see if the number of failures was more than we could expect
         // by chance. first we'll calculate some parameters about the Bloom filter using the same logic as the filter
